Generate Messages Azure Function App/requirements.txt
azure-functions
azure-storage-blob
xmltodict
requests
jsonschema